requirements.txt
# Backend - FastAPI Voice Recognition System (Hardware Limitado)

# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Procesamiento de audio y reconocimiento de voz
pyannote.audio>=3.0.0
librosa>=0.10.0

# Análisis numérico y científico (versiones ligeras)
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualización (opcional para hardware limitado)
# matplotlib>=3.7.0

# Procesamiento de archivos
soundfile>=0.12.0

# Validación de datos
pydantic>=2.0.0

# Utilidades
python-multipart>=0.0.6
psutil>=5.8.0  # Para monitoreo de recursos 